WIP - structure of FitsFileLoaderV2 now works
The core elements of FitsFileLoaderV2 now go through without
any unexpected errors or rejected promises. However, the
parseSingleExtensionFitsFile function is returning no data. Looking
at the console, it would appear that the sub-functions processing
the data are being called *after* the master parsing function
completes; or, in other words, it looks like the main function is
running off to completion before the data is actually ready.