WIP - 6dF table read-in
Delving further into fits.js, it certainly looks like the
developers of that code have written a lot of stuff into the
Tabular parent class which really belongs to the BinaryTable
child. A lot of the assumptions functions in Tabular are making
depend on properties and functions of BinaryTable, which seems
to be breaking the Table class. More work on this required.